dsds

+++ lab5/src/lab5/data/Time.java
@@ -6,7 +6,7 @@ package lab5.data;
  *
  */
 public class Time {
-	private long elapsedTime;
+	private double elapsedTime;
 	
 	/**
 	 * Makes time object
